+++ src/oihana/magento/schema/constants/traits/ProductTrait.php
@@ -5,6 +5,7 @@ namespace oihana\magento\schema\constants\traits ;
 trait ProductTrait
 {
     public const string CANONICAL_URL         = 'canonical_url' ;
+    public const string ENTITY_ID             = 'entity_id' ;
     public const string MEDIA_GALLERY         = 'media_gallery' ;
     public const string MEDIA_GALLERY_ENTRIES = 'media_gallery_entries' ;
     public const string PRICE                 = 'price' ;